requirements.txt
openai>=1.14.0
python-dotenv>=1.0.0
Pillow>=10.0.0
click>=8.1.0
Django>=4.2
django-cors-headers>=4.3.0
